Visualization now contains all the languages in the text but they're not in the right order.
Changed Germanic to Proto-Germanic to make it more historically accurate.

Added Old Norse and Proto-Norse to languageHierarchy.json to specify the parents although they're not in the text (needed for visualization).

Rewrote converter.py. Ensured paths will always be correct (dynamic). Uses both the output from the parser and child-parent mapping to create a JSON suitable for the visualization.